# Upgraded to CEL-0.4.0-beta.1 and added minor memory optimization