--- a/go.mod
+++ b/go.mod
@@ -4,8 +4,12 @@ go 1.12
 
 require (
 	github.com/golang/protobuf v1.3.2
-	github.com/google/cel-go v0.4.0-alpha.2
+	github.com/google/cel-go v0.4.0-beta.1
+	github.com/google/go-cmp v0.4.0 // indirect
+	golang.org/x/net v0.0.0-20191002035440-2ec189313ef0 // indirect
+	golang.org/x/sys v0.0.0-20191204072324-ce4227a45e2e // indirect
 	google.golang.org/genproto v0.0.0-20191007204434-a023cd5227bd
+	google.golang.org/grpc v1.27.0 // indirect
 	gopkg.in/yaml.v2 v2.2.4 // indirect
 	gopkg.in/yaml.v3 v3.0.0-20191010095647-fc94e3f71652
 )