Upgraded the CEL core library dependency to take advantage of recent performance optimizations.
Also added a benchmark test and made a minor memory optimization

```
benchmark                                                                       old ns/op     new ns/op     delta
BenchmarkEnforcer/required_label_violation-8                                    9053          9606          +6.11%
BenchmarkEnforcer/timed_contract_valid-8                                        4408          3294          -25.27%
BenchmarkEnforcer/timed_contract_no_match-8                                     1400          942           -32.71%
BenchmarkEnforcer/timed_contract_invalid-8                                      4921          3504          -28.79%
BenchmarkEnforcer/sensitive_data_prefix_same_location-8                         1113          1175          +5.57%
BenchmarkEnforcer/sensitive_data_prefix_diff_location-8                         3541          3327          -6.04%
BenchmarkEnforcer/sensitive_data_label_same_location-8                          1072          1161          +8.30%
BenchmarkEnforcer/sensitive_data_label_diff_location-8                          7810          7024          -10.06%
BenchmarkEnforcer/sensitive_data_diff_location_not_sensitive-8                  7114          6871          -3.42%
BenchmarkEnforcer/restricted_destinations_valid_location-8                      7425          5049          -32.00%
BenchmarkEnforcer/restricted_destinations_restricted_location_us_national-8     5048          3466          -31.34%
BenchmarkEnforcer/restricted_destinations_restricted_location_ir_national-8     2966          2113          -28.76%
BenchmarkEnforcer/restricted_destinations_valic_location_ir_label-8             6925          5048          -27.10%
BenchmarkEnforcer/allowed_resource_types_denied_request-8                       5123          4297          -16.12%

benchmark                                                                       old allocs     new allocs     delta
BenchmarkEnforcer/required_label_violation-8                                    72             70             -2.78%
BenchmarkEnforcer/timed_contract_valid-8                                        30             12             -60.00%
BenchmarkEnforcer/timed_contract_no_match-8                                     15             8              -46.67%
BenchmarkEnforcer/timed_contract_invalid-8                                      32             14             -56.25%
BenchmarkEnforcer/sensitive_data_prefix_same_location-8                         10             7              -30.00%
BenchmarkEnforcer/sensitive_data_prefix_diff_location-8                         28             20             -28.57%
BenchmarkEnforcer/sensitive_data_label_same_location-8                          10             7              -30.00%
BenchmarkEnforcer/sensitive_data_label_diff_location-8                          66             51             -22.73%
BenchmarkEnforcer/sensitive_data_diff_location_not_sensitive-8                  66             51             -22.73%
BenchmarkEnforcer/restricted_destinations_valid_location-8                      62             38             -38.71%
BenchmarkEnforcer/restricted_destinations_restricted_location_us_national-8     41             21             -48.78%
BenchmarkEnforcer/restricted_destinations_restricted_location_ir_national-8     26             12             -53.85%
BenchmarkEnforcer/restricted_destinations_valic_location_ir_label-8             58             34             -41.38%
BenchmarkEnforcer/allowed_resource_types_denied_request-8                       46             28             -39.13%

benchmark                                                                       old bytes     new bytes     delta
BenchmarkEnforcer/required_label_violation-8                                    3350          3082          -8.00%
BenchmarkEnforcer/timed_contract_valid-8                                        1302          623           -52.15%
BenchmarkEnforcer/timed_contract_no_match-8                                     679           485           -28.57%
BenchmarkEnforcer/timed_contract_invalid-8                                      1335          659           -50.64%
BenchmarkEnforcer/sensitive_data_prefix_same_location-8                         839           472           -43.74%
BenchmarkEnforcer/sensitive_data_prefix_diff_location-8                         1249          767           -38.59%
BenchmarkEnforcer/sensitive_data_label_same_location-8                          839           472           -43.74%
BenchmarkEnforcer/sensitive_data_label_diff_location-8                          2134          1559          -26.94%
BenchmarkEnforcer/sensitive_data_diff_location_not_sensitive-8                  2129          1549          -27.24%
BenchmarkEnforcer/restricted_destinations_valid_location-8                      1966          1084          -44.86%
BenchmarkEnforcer/restricted_destinations_restricted_location_us_national-8     1600          815           -49.06%
BenchmarkEnforcer/restricted_destinations_restricted_location_ir_national-8     1284          617           -51.95%
BenchmarkEnforcer/restricted_destinations_valic_location_ir_label-8             1935          1043          -46.10%
BenchmarkEnforcer/allowed_resource_types_denied_request-8                       2348          1655          -29.51%
```